refactor(gui): decompose monolithic _gui_impl.py into distinct modules
